role feature is done

+++ app/Services/Auth/Role/RoleService.php
@@ -11,7 +11,11 @@ namespace App\Services\Auth\Role;
 
 use App\Models\Auth\Role;
 use App\Models\Auth\User;
+use Cartalyst\Sentinel\Laravel\Facades\Sentinel;
 use Illuminate\Pagination\Paginator;
+use Illuminate\Support\Str;
+use Yajra\DataTables\Facades\DataTables;
+use Illuminate\Support\Facades\DB;
 
 /**
  * Class RoleService
@@ -61,7 +65,7 @@ class RoleService implements RoleServiceContract
     {
         // TODO: Implement getData() method.
         $select = [
-            'name', 'slug', 'created_by', 'updated_by', 'created_at', 'updated_at'
+            'id','name', 'slug', 'created_by', 'updated_by', 'created_at', 'updated_at'
         ];
 
         $dataDb = $this->model::select($select);
@@ -77,43 +81,40 @@ class RoleService implements RoleServiceContract
     public function store($request)
     {
         // TODO: Implement store() method.
+        # retrieve
+        $userDb = Sentinel::getUser()->name;
+
         DB::beginTransaction();
         try {
-            # logic to insert to DB
-            $userDb = Sentinel::getUser();
-            $email = $request->email;
+            # insert to role
+            $data             = new Role();
+            $data->name       = $request->name;
+            $data->slug       = Str::slug( $request->name );
+            $data->created_by = $userDb;
+            $data->updated_by = $userDb;
 
-            #inserto into DB
-            $data = [
-                'name'       => $request->name,
-                'phone'      => $request->phone,
-                'email'      => strtolower( $email ),
-                'password'   => $request->password,
-                'created_by' => $userDb->name,
-                'updated_by' => $userDb->name
-            ];
+            /**
+             *  Permission Here
+             */
+            $permissions       = collect( json_decode( $this->permissions( $request ) ) )->toArray();
+            $data->permissions = $permissions;
 
-            #create new user
-            $user = Sentinel::registerAndActivate( $data );
-
-            # attach user into role
-            $role = Sentinel::findRoleById( $request->role );
-            $role->users()->attach( $user );
+            $data->save();
 
             # commit to insert to DB
             DB::commit();
 
             # return to controller
-            return $user;
+            return $data;
         } catch (\Exception $exception) {
             #rollback to begin (not insert to DB)
             DB::rollBack();
 
             #Dump
-            dd('Message: '.$exception->getMessage() . ' Line: ' . $exception->getLine()  . ' Code: ' . $exception->getCode());
+            //dd('Message: '.$exception->getMessage() . ' Line: ' . $exception->getLine()  . ' Code: ' . $exception->getCode());
 
             #return getCode Error
-            return $e->getCode();
+            return $exception->getCode();
         }
     }
 
@@ -125,48 +126,40 @@ class RoleService implements RoleServiceContract
     public function update(int $id, $request)
     {
         // TODO: Implement update() method.
-        $user = Sentinel::findById( $id );
 
         DB::beginTransaction();
         try {
+            $dataDb = Role::find( $id );
 
-            $oldRole = Sentinel::findRoleById( $user->roles[0]->id ?? null );
+            if ( empty( $dataDb ) ) {
+                Session::flash( 'failed', 'Access Denied' );
 
-            $credentials = [
-                'name'       => $request->name,
-                'phone'      => $request->phone,
-                'updated_by' => $user->name
-            ];
-
-            #If User Input Password
-            if ( $request->password ) {
-                $credentials['password'] = $request->password;
+                return true;
             }
 
-            #Valid User For Update
-            $role = Sentinel::findRoleById( $request->role );
+            $userDb             = Sentinel::getUser()->name;
+            $dataDb->name       = $request->name;
+            #many function use slug as identifier, not recommend for change it
+            #$dataDb->slug       = Str::slug( $request->name );
+            $dataDb->updated_by = $userDb;
 
-            if ( $oldRole ) {
-                #Remove a user from a role.
-                $oldRole->users()
-                    ->detach( $user );
-            }
+            /**
+             *  Permission Here
+             */
+            $permissions         = collect( json_decode( $this->permissions( $request ) ) )->toArray();
+            $dataDb->permissions = $permissions;
+            $dataDb->save();
 
-            #Assign a user to a role.
-            $role->users()
-                ->attach( $user );
 
-            #Update User
-            $updateIntoDb = Sentinel::update( $user, $credentials );
             DB::commit();
 
-            return $updateIntoDb;
+            return $dataDb;
 
         } catch ( \Exception $exception ) {
 
             DB::rollBack();
 
-            dd('Message: '.$exception->getMessage() . ' Line: ' . $exception->getLine()  . ' Code: ' . $exception->getCode());
+            //dd('Message: '.$exception->getMessage() . ' Line: ' . $exception->getLine()  . ' Code: ' . $exception->getCode());
 
             return $exception->getCode();
         }
@@ -176,10 +169,23 @@ class RoleService implements RoleServiceContract
      * @param int $id
      * @return mixed|void
      */
-    public function destroy($user)
+    public function destroy($id)
     {
         // TODO: Implement destroy() method.
-        return $user->delete();
+        $userDb = Sentinel::getUser();
+        $dataDb = Sentinel::findRoleById( $id );
+
+        #if user root || data is empty || restricted role is can't delete
+        if ( empty( $dataDb ) || $this->restricted($dataDb->slug) ) {
+
+            return false;
+        }
+
+        $dataDb->users()
+            ->detach( $userDb );
+        $dataDb->delete();
+
+        return $dataDb;
     }
 
     /**
@@ -203,63 +209,36 @@ class RoleService implements RoleServiceContract
 
         return DataTables::eloquent($query)
             ->editColumn('created_at', function($dataDb) {
-                return $dataDb->created_at->format('d M Y H:i:s');
+                return $dataDb->created_at->format('Y-m-d H:i:s');
             })
             ->editColumn('updated_at', function($dataDb) {
-                return $dataDb->updated_at->format('d M Y H:i:s');
+                return $dataDb->updated_at->format('Y-m-d H:i:s');
             })
-            ->addColumn('status', function($dataDb) {
-                if ($dataDb->activations->isNotEmpty()) {
-                    if ($dataDb->activations[0]->completed == 1) {
-                        return '<a href="#" data-message="Are you sure you want to deactivate the user \''.$dataDb->name.'\' ?"
-                    data-href="'.route('user.status', $dataDb->id).'"
-                    data-title="Deactive This User"
-                    data-title-modal="Deactive This User"
-                    data-method="PUT"
-                    data-toggle="modal"
-                    data-target="#delete"
-                    title="Deactive this user"><span class="kt-badge  kt-badge--success kt-badge--inline kt-badge--pill">Active</span></a>';
-                    }
-                }
-                return '<a href="#" data-message="Are you sure you want to activate the user \''.$dataDb->name.'\' ?"
-                    data-href="'.route('user.status', $dataDb->id).'"
-                    data-title="Activate This User"
-                    data-title-modal="Activate This User"
-                    data-method="PUT"
-                    data-toggle="modal"
-                    data-target="#delete"
-                    title="Activated this user"><span class="kt-badge  kt-badge--danger kt-badge--inline kt-badge--pill">Inactive</span></a>';
+           ->addColumn('action', function ($dataDb){
+               $btnShow = '';
+               $btnEdit = '';
+               $btnDelete = '';
 
-            })
-            ->addColumn('role', function($dataDb) {
-                if ($dataDb->roles->isNotEmpty()) {
-                    return implode(', ', collect( $dataDb->roles )
-                        ->pluck('name')
-                        ->all() );
-                }
-            })
-//            ->addColumn(
-//                'checkbox',
-//                function ($dataDb) {
-//                    return '<label class="kt-checkbox kt-checkbox--single kt-checkbox--solid kt-checkbox--brand"><input type="checkbox" value="'.$dataDb->id.'" class="kt-checkable"><span></span></label>'; //$dataDb->id;
-//                }
-//            )
-            ->addColumn('action', function ($dataDb){
-                $btnShow = '<a href="'.route('user.show', $dataDb->id).'" id="tooltip" title="'.trans('global.show').'">
+               if (Sentinel::inRole('root') || Sentinel::hasAccess(['role.show'])) {
+                   $btnShow = '<a href="'.route('roles.show', $dataDb->id).'" id="tooltip" title="Show">
                             <span class="label label-primary label-sm">
                                 <i class="fa fa-arrows-alt"></i>
                                 </span>
                             </a>';
+               }
 
-                $btnEdit = '<a href="'.route('user.edit', $dataDb->id).'"
+               if (Sentinel::inRole('root') || Sentinel::hasAccess(['role.edit'])) {
+                   $btnEdit = '<a href="'.route('roles.edit', $dataDb->id).'"
                             data-tooltip-custom="tooltip" data-placement="left" title="Update It" >
                             <i class="flaticon-edit kt-font-brand"></i>
                             </a>
                             ';
+               }
 
-                $btnDelete = '<a href="#"
+               if (Sentinel::inRole('root') || Sentinel::hasAccess(['role.destroy'])) {
+                   $btnDelete = '<a href="#"
                                 data-message="Are u sure for delete this? '.$dataDb->name.'"
-                                data-href="'.route('user.destroy', $dataDb->id).'"
+                                data-href="'.route('roles.destroy', $dataDb->id).'"
                                 data-method="DELETE"
                                 title="delete"
                                 data-original-title="Delete It?"
@@ -269,47 +248,11 @@ class RoleService implements RoleServiceContract
                                 data-target="#delete">
                                 <i class="flaticon-delete kt-font-danger"></i>
                     </a>';
+               }
                 return $btnEdit . $btnDelete;
             })
-            ->rawColumns(['action', 'status'])
+            ->rawColumns(['action'])
             ->make(true);
-    }
-
-    /**
-     * @param int $id
-     * @return bool|mixed
-     */
-    public function status(int $id)
-    {
-        // TODO: Implement status() method.
-        $user = Sentinel::findById( $id );
-
-        $activation = Activation::completed( $user );
-
-        if ( $activation !== false ) {
-            #Deactivated This Activation
-            if ( $user->id === Sentinel::getUser()->id ) {
-                return false;
-            }
-
-            #Remove this account
-            Activation::remove( $user );
-            return true;
-        }
-
-        #Deactivated This Activation
-        if ( $user->id === Sentinel::getUser()->id ) {
-
-            return false;
-        }
-
-        #Get Activation Code
-        $activationCreate = Activation::create( $user );
-
-        #Activate this account
-        Activation::complete( $user, $activationCreate->code );
-
-        return true;
     }
 
     /**
@@ -326,8 +269,26 @@ class RoleService implements RoleServiceContract
             return $page;
         });
 
-        $dataDb = User::select(['id', 'name as text', 'phone'])->where('name', 'LIKE', '%' . $request->term . '%')->where('type', 'user')->orderBy('name')->paginate($perPage);
+        $dataDb = Role::select(['id', 'name as text'])->where('name', 'LIKE', '%' . $request->term . '%')->orderBy('name')->paginate($perPage);
 
         return $dataDb;
+    }
+
+    private function permissions( $request ) {
+
+        //Dashboard
+        $permissions['dashboard'] = true;
+
+        $request = $request->except( array( '_token', 'name', '_method', 'previousUrl' ) );
+
+        foreach ( $request as $key => $value ) {
+            $permissions[ preg_replace( '/_([^_]*)$/', '.\1', $key ) ] = true;
+        }
+
+        return json_encode( $permissions );
+    }
+
+    private function restricted($slug){
+        return $slug == 'root' || $slug == 'member' || $slug == 'sales';
     }
 }
